move xml compare into python-tools
